refactor: 구조 변경
1. 상대방 역할을 하는 객체는 불필요하다고 판단하였고, 상대방의 number는 심판객체가 스스로 알고있도록 수정
2. 심판객체의 judgeBallOrStrike로 사용자의 number만 요청하여 ball, strike개수를 응답받도록 수정
3. 심판 객체가 사용자 객체를 알고있어 set을 통해 수정하는 부분을 제거하여 결합도를 낮추도록 수정
