+++ src/App.js
@@ -1,7 +1,6 @@
 import InputView from './View/InputView.js';
 import OutputView from './View/OutputView.js';
 
-import Opponent from './Opponent.js';
 import Player from './Player.js';
 import Refree from './Refree.js';
 import Validation from './Validation.js';
@@ -16,22 +15,19 @@ class App {
   async play() {
     const inputNumber = await InputView.getUserNumber();
     Validation.validateBaseballNumber(inputNumber);
+    this.#refree = new Refree();
 
-    const playerNumber = new Player(inputNumber).numberArray;
-    const opponentNumber = new Opponent().opponentNumber;
-
-    this.#refree = new Refree(playerNumber, opponentNumber);
-    await this.startGame();
+    await this.startGameRound(new Player(inputNumber).numberArray);
   }
 
-  async startGame() {
-    const { ball, strike } = this.#refree.playGame();
+  async startGameRound(playerNumber) {
+    const { ball, strike } = this.#refree.judgeBallOrStrike(playerNumber);
     OutputView.printResult(ball, strike);
     await this.checkSuccess(strike);
   }
 
   async checkSuccess(strike) {
-    strike === 3 ? this.confirmGame() : await this.continueGame();
+    strike === 3 ? await this.confirmGame() : await this.continueGame();
   }
 
   async confirmGame() {
@@ -39,15 +35,14 @@ class App {
     const inputNumber = await InputView.confirmContinue();
     Validation.validateConfirmNumber(inputNumber);
 
-    if (Number(inputNumber) === 1) this.play();
+    if (Number(inputNumber) === 1) await this.play();
     if (Number(inputNumber) === 2) return;
   }
 
   async continueGame() {
     const inputNumber = await InputView.getUserNumber();
     Validation.validateBaseballNumber(inputNumber);
-    this.#refree.changePlayerNumber(inputNumber);
-    this.startGame();
+    await this.startGameRound(new Player(inputNumber).numberArray);
   }
 }
